fix bugs in example codec
